Add files via upload
Requires requests and BeautifulSoup.
Make sure to make a SBS_DATA.txt file in the same folder (I cooked this up in a few minutes. Didn't have time to tinker with file permissions).
Ideas on how to use this: Use this and as Excel macro that calls this and then imports the SBS_DATA.txt file. Do cool stuff to the data and hope you can come out with your life savings intact.